reorganizing files + fixing constraints
