+++ SocialMap/Source/View/General/Annotation/Components/UserImageAnnotation.swift
