+++ SocialMap/Source/View/General/CustomGestureRecognizer.swift
@@ -4,7 +4,6 @@ import UIKit
 import MapKit
 
 class CustomGestureRecognizer : UILongPressGestureRecognizer {
-    
     var longPressCallback: ((_ location: Location, _ mapView: MKMapView) -> Void)?
     var oneClickCallback: ((_ point: CGPoint, _ mapView: MKMapView) -> Void)?
 }
